Keep superuser_required with the other other auth helpers
