+++ jobserver/roles.py
@@ -1,3 +1,8 @@
+from functools import wraps
+
+from django.contrib import messages
+from django.shortcuts import redirect
+
 from .github import is_member_of_org
 
 
@@ -13,3 +18,21 @@ def can_run_jobs(user):
         return False
 
     return is_member_of_org("opensafely", user.username)
+
+
+def superuser_required(f):
+    """
+    Decorator for views which require a Superuser
+
+    User.is_superuser implies the User is authenticated.
+    """
+
+    @wraps(f)
+    def wrapper(request, *args, **kwargs):
+        if request.user.is_superuser:
+            return f(request, *args, **kwargs)
+
+        messages.error(request, "Only admins can view Backends.")
+        return redirect("/")
+
+    return wrapper
